// src/api/models/couponInfoListItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Allta Admin
 * Allta Admin 서버 문서입니다.
 * OpenAPI spec version: 1.0
 */

export interface CouponInfoListItem {
  /** ID */
  id: string;
  /** 쿠폰 이름 */
  name: string;
  /** 쿠폰 종류 */
  type: string;
  /**
   * 이용권 종류
   * @nullable
   */
  passType?: string | null;
  /**
   * 서비스 종류
   * @nullable
   */
  serviceType?: string | null;
  /**
   * 등록 코드
   * @nullable
   */
  code?: string | null;
  /** 할인 종류 */
  discountType: string;
  /** 할인 금액 */
  discountValue: number;
  /**
   * 쿠폰 등록 시작일
   * @nullable
   */
  startDate?: string | null;
  /**
   * 쿠폰 등록 종료일
   * @nullable
   */
  endDate: string | null;
  /** 쿠폰 사용기간 */
  validDays: number;
  /**
   * 발급 수량
   * @nullable
   */
  maxQuantity?: number | null;
  /** 생성일 */
  createdAt: string;
  /** 활성화 여부 */
  isActive: boolean;
  /** 발급 중단 여부 */
  isExhausted: boolean;
  /** 중복 수령 가능 여부 */
  isMultiple: boolean;
}